requirements.txt
dfss
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
transformers>=4.39.1
tpu_perf-1.2.35-py3-none-manylinux2014_aarch64.whl